Delete README.md
I am a banker who is passionate about computer programming, and I am going to use this forum to learn programming. I hot into the banking industry because it was the cheapest and easiest thing to for me to do after high school. Going to school to pursue further studies was a non-starter as the resources were not permitting. A chance to learn came again as one of the local universities offered evening degree programmes. Again the night school did not have computer science or related programs.  So I found myself going for a commercial course as I was already working a banking environment. However this was not my passion, but thanks to this program because I can finally pursue my dream.